split service manager into unit
